Feature 20150923

+++ tiger/src/main/java/com/dianping/tiger/dispatch/DispatchMultiService.java
@@ -32,6 +32,6 @@ public interface DispatchMultiService extends DispatchTaskService {
 	 * @return
 	 */
 	public List<DispatchTaskEntity> findDispatchTasksWithLimitByBackFetch(
-			String handler, List<Integer> nodeList, int limit, int taskId);
+			String handler, List<Integer> nodeList, int limit, long taskId);
 
 }
